split api install integration test
