+++ api/integration/assets/embed.go
@@ -1,0 +1,6 @@
+package assets
+
+import _ "embed"
+
+//go:embed license.yaml
+var LicenseData []byte
