+++ api/integration/auth/controller_test.go
@@ -1,4 +1,4 @@
-package integration
+package auth
 
 import (
 	"bytes"
